+++ Cargo.toml
@@ -22,6 +22,8 @@ features = ["derive"]
 version = "1.0.82"
 features = ["arbitrary_precision"]
 
+[patch.crates-io]
+themelio-nodeprot={path="../themelio-nodeprot"}
 
 [dependencies]
 acidjson = "0.1.2"
@@ -53,9 +55,10 @@ clap = { version = "3.2.12", features = ["derive"] }
 smol-timeout = "0.6.0"
 smolscale= "0.3.29"
 tap = "1.0.1"
-themelio-bootstrap = "0.6.0"
-themelio-nodeprot = "0.11.5"
-# themelio-nodeprot = { path = "../themelio-nodeprot" }
+themelio-bootstrap = "0.7.1"
+themelio-nodeprot = "0.12.0"
+# melnet = "0.2.27"
+
 themelio-stf = "0.11.2"
 themelio-structs = "0.2.6"
 tide = "0.16.0"
@@ -66,12 +69,12 @@ serde_yaml = "0.8.26"
 display_json = "0.2.1"
 terminal_size = "0.2.1"
 nanorpc = "0.1.7"
-thiserror = "1.0.36"
-melnet = "0.2.27"
 async-trait = "0.1.57"
 route-recognizer = "0.3.1"
 futures = "0.3.24"
 melwalletd-prot = {path = "../melwalletd-prot"}
+
+[dev-dependencies]
 
 # prevent key derivation from being ridiculously slow
 [profile.dev.package.rust-argon2]
